--- nexus/internal/uploader/uploader_azure.go
@@ -1,1 +1,0 @@
-package uploader
