+++ nexus/internal/uploader/uploader_default.go
@@ -1,0 +1,232 @@
+package uploader
+
+import (
+	"fmt"
+	"io"
+	"log"
+	"log/slog"
+	"net/http"
+	"os"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/hashicorp/go-retryablehttp"
+	"github.com/wandb/wandb/nexus/pkg/observability"
+)
+
+const (
+	//ChunkSize       = 1024 * 1024 // 1MB chunks
+	ChunkSize       = 200
+	FileParameter   = "file"
+	ContentRangeFmt = "bytes %d-%d/%d"
+)
+
+type RetryClientOption func(rc *retryablehttp.Client)
+
+func WithRetryClientLogger(logger *observability.NexusLogger) RetryClientOption {
+	return func(rc *retryablehttp.Client) {
+		rc.Logger = slog.NewLogLogger(logger.Logger.Handler(), slog.LevelDebug)
+	}
+}
+
+func WithRetryClientRetryMax(retryMax int) RetryClientOption {
+	return func(rc *retryablehttp.Client) {
+		rc.RetryMax = retryMax
+	}
+}
+
+func WithRetryClientRetryWaitMin(retryWaitMin time.Duration) RetryClientOption {
+	return func(rc *retryablehttp.Client) {
+		rc.RetryWaitMin = retryWaitMin
+	}
+}
+
+func WithRetryClientRetryWaitMax(retryWaitMax time.Duration) RetryClientOption {
+	return func(rc *retryablehttp.Client) {
+		rc.RetryWaitMax = retryWaitMax
+	}
+}
+
+func NewRetryClient(opts ...RetryClientOption) *retryablehttp.Client {
+	retryClient := retryablehttp.NewClient()
+
+	for _, opt := range opts {
+		opt(retryClient)
+	}
+	return retryClient
+}
+
+// DefaultUploader uploads files to the server
+type DefaultUploader struct {
+	// client is the HTTP client for the uploader
+	client *retryablehttp.Client
+
+	// logger is the logger for the uploader
+	logger *observability.NexusLogger
+}
+
+// NewDefaultUploader creates a new uploader
+func NewDefaultUploader(logger *observability.NexusLogger) *DefaultUploader {
+	retryClient := NewRetryClient(
+		WithRetryClientLogger(logger),
+		WithRetryClientRetryMax(10),                 // todo: make this configurable
+		WithRetryClientRetryWaitMin(1*time.Second),  // todo: make this configurable
+		WithRetryClientRetryWaitMax(60*time.Second), // todo: make this configurable
+	)
+
+	uploader := &DefaultUploader{
+		client: retryClient,
+		logger: logger,
+	}
+	return uploader
+}
+
+// Upload uploads a file to the server
+func (u *DefaultUploader) Upload(task *UploadTask) error {
+	u.logger.Debug("custom uploader: uploading file", "path", task.Path, "url", task.Url)
+	// open the file for reading and defer closing it
+	file, err := os.Open(task.Path)
+	if err != nil {
+		task.outstandingDone()
+		return err
+	}
+	defer func(file *os.File) {
+		err := file.Close()
+		if err != nil {
+			u.logger.CaptureError("uploader: error closing file", err, "path", task.Path)
+		}
+	}(file)
+
+	fileStat, err := file.Stat()
+	fileSize := fileStat.Size()
+
+	chunk := make([]byte, ChunkSize)
+	var start, end int64
+
+	// Initiate the resumable upload
+	req, err := retryablehttp.NewRequest(http.MethodPost, task.Url, nil)
+	if err != nil {
+		task.outstandingDone()
+		return err
+	}
+	req.Header.Set("Content-Length", "0")
+	//req.Header.Set("Content-Type", "")
+	// x-goog-resumable
+	req.Header.Set("x-goog-resumable", "start")
+
+	resp, err := u.client.Do(req)
+	if err != nil {
+		task.outstandingDone()
+		return err
+	}
+	fmt.Println(resp.StatusCode, resp.Header, resp)
+
+	for {
+		n, err := file.Read(chunk)
+		if err != nil {
+			if err == io.EOF {
+				break
+			}
+			u.logger.CaptureError("uploader: error reading file chunk", err, "path", task.Path)
+			task.outstandingDone()
+			return err
+		}
+
+		end = start + int64(n) - 1
+		contentRange := fmt.Sprintf(ContentRangeFmt, start, end, fileSize)
+
+		//body := &bytes.Buffer{}
+		//writer := multipart.NewWriter(body)
+		//part, err := writer.CreateFormFile(FileParameter, fileStat.Name())
+		//if err != nil {
+		//	u.logger.CaptureError("uploader: error creating form file", err, "path", task.Path)
+		//	task.outstandingDone()
+		//	return err
+		//}
+		//
+		//if _, err := part.Write(chunk[:n]); err != nil {
+		//	u.logger.CaptureError("uploader: error writing chunk data", err, "path", task.Path)
+		//	task.outstandingDone()
+		//	return err
+		//}
+		//
+		//if err := writer.Close(); err != nil {
+		//	u.logger.CaptureError("uploader: error closing multipart writer", err, "path", task.Path)
+		//	task.outstandingDone()
+		//	return err
+		//}
+
+		//fmt.Println(body.String())
+		fmt.Println(contentRange)
+
+		req, err := retryablehttp.NewRequest(http.MethodPut, task.Url, chunk[:n])
+		if err != nil {
+			log.Fatalf("Could not create request: %v", err)
+		}
+
+		for _, header := range task.Headers {
+			parts := strings.Split(header, ":")
+			req.Header.Set(parts[0], parts[1])
+		}
+
+		//req.Header.Set("Content-Type", writer.FormDataContentType())
+		req.Header.Set("Content-Range", contentRange)
+		req.Header.Set("Content-Length", strconv.Itoa(n))
+		//req.Header.Set("Content-Type", "application/octet-stream")
+
+		resp, err := u.client.Do(req)
+		if err != nil {
+			u.logger.CaptureError("uploader: error sending request", err, "path", task.Path)
+			task.outstandingDone()
+			return err
+		}
+		// todo: handle non-200 status codes
+		fmt.Println(resp.StatusCode)
+		//if resp.StatusCode != 200 {
+		//	err = fmt.Errorf("uploader: received non-200 status code: %d", resp.StatusCode)
+		//	u.logger.CaptureError(
+		//		"uploader: received non-200 status code",
+		//		err,
+		//		"path", task.Path, "status", resp.StatusCode, "status_text", resp.Status)
+		//	task.outstandingDone()
+		//	return err
+		//}
+
+		start += int64(n)
+	}
+
+	req, err = retryablehttp.NewRequest(http.MethodPut, task.Url, nil)
+	if err != nil {
+		task.outstandingDone()
+		return err
+	}
+	req.Header.Set("Content-Length", "0")
+	req.Header.Set("Content-Type", "")
+	// x-goog-resumable
+	req.Header.Set("x-goog-resumable", "final")
+
+	//req, err := retryablehttp.NewRequest(
+	//	http.MethodPut,
+	//	task.Url,
+	//	file,
+	//)
+	//
+	//for _, header := range task.Headers {
+	//	parts := strings.Split(header, ":")
+	//	req.Header.Set(parts[0], parts[1])
+	//}
+	//
+	//if err != nil {
+	//	task.outstandingDone()
+	//	return err
+	//}
+	//
+	//if _, err = u.client.Do(req); err != nil {
+	//	task.outstandingDone()
+	//	return err
+	//}
+
+	task.outstandingDone()
+	return nil
+}
