--- nexus/internal/uploader/uploader_s3.go
@@ -1,1 +1,0 @@
-package uploader
